fix(frontend): align protocol console with regenerated program surface
Whole frontend typechecks clean and test:node passes 315/315.

Surface-alignment fixes:
- Drop the spurious writable flag on the signer authority in the rotate/
  accept/cancel governance builders so the AccountMeta matches the canonical
  signer-only surface (mirrors buildSetProtocolEmergencyPauseTx).
- Add buildAuthorizeClaimRecipientTx for parity with the rest of the claims
  builder family (authority, protocol_governance, member_position, claim_case;
  args delegate_recipient).
- Repoint the governance-workbench role->action map at canonical instruction
  names: governance_execute->rotate_protocol_governance_authority,
  rail_posture->configure_reserve_asset_rail,
  update_policy_series_controls->version_policy_series,
  create_funding_line->open_funding_line,
  fund_plan_liabilities->fund_sponsor_budget. Clear pool_sentinel actions
  since the program grants the sentinel no direct instruction authority.

Folds in the in-progress protocol model reorganization that protocol.ts now
depends on (reserve-readiness/waterfall helpers extracted to
frontend/lib/protocol/reserve-readiness.ts; buildUpdateLiquidityPoolControlsTx
added) so the committed state builds.
